requirements.txt
streamlit==1.30.0
openai==1.3.8
gspread==5.12.0
google-auth==2.23.4